requirements.txt
flask
python-docx
spacy
spacy-transformers==1.2.2
Flask-Cors==3.0.10
pandas==1.5.3
plotly==5.13.1